+++ src/ufaya/firewall/__init__.py
@@ -1,12 +1,14 @@
 """Firewall abstractions."""
 
-from ufaya.models.firewall_rule import FirewallRule
+from typing import Any
+
+from ufaya.models.firewall_rule import FirewallRuleRecord
 
 from ..services.device_factory import get_firewall_driver
 from .base import FirewallDriver
 
 
-def get_rules(vendor: str, **kwargs: str) -> list[FirewallRule]:
+def get_rules(vendor: str, **kwargs: Any) -> list[FirewallRuleRecord]:
     driver = get_firewall_driver(vendor, **kwargs)
     return driver.get_rules()